Using buffered channels to increase the throughput or make sure the websockets don't block until we update the items in our datastore
Also fixing the bug where main.BuildTimestamp was not properly set as part of the build. Turns out only string values can be set from -X as part of the linking process and not other value / types as far as I can see. Also started reusing the fastjson.Parser to make it more efficient in memory management to reduce the # of GC cycles over time